Make installable
Install with `pip install -e .` from the virtual environment.

Run with `musicmanager <ARGS>`.

This enables us to run the tool from the command line directly and get source changes instantly.

The package name was renamed from "music_manager" to "musicmanager" to remove the underscore, which PEP8 indicates we should not have.

The main module was renamed from "music_manager" to "core" to deconflict with the package name.