Add script to generate an access token
The Spotify developer console is less easy to access now. This script makes it easier and quicker to generate a token.

We use the Client Credentials Flow: https://developer.spotify.com/documentation/web-api/tutorials/client-credentials-flow

The credentials are retrieved from environment variables.

This is less sustainable long term, but it avoids user authentication for now which requires a browser.